mgr/nfs: create new module for export utils
(cherry picked from commit d064306dd3afcbab620c5489654359bb2f911c3d)
